+++ packages/abstract-signer/lib.esm/index.js
@@ -15,6 +15,11 @@ const logger = new Logger(version);
 const allowedTransactionKeys = [
     "chainId", "data", "from", "gasLimit", "gasPrice", "nonce", "to", "value"
 ];
+const forwardErrors = [
+    Logger.errors.INSUFFICIENT_FUNDS,
+    Logger.errors.NONCE_EXPIRED,
+    Logger.errors.REPLACEMENT_UNDERPRICED,
+];
 // Sub-Class Notes:
 //  - A Signer MUST always make sure, that if present, the "from" field
 //    matches the Signer, before sending or signing a transaction
@@ -137,6 +142,9 @@ export class Signer {
             }
             if (tx.gasLimit == null) {
                 tx.gasLimit = this.estimateGas(tx).catch((error) => {
+                    if (forwardErrors.indexOf(error.code) >= 0) {
+                        throw error;
+                    }
                     return logger.throwError("cannot estimate gas; transaction may fail or may require manual gas limit", Logger.errors.UNPREDICTABLE_GAS_LIMIT, {
                         error: error,
                         tx: tx