+++ packages/abstract-signer/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "abstract-signer/5.0.4";
+export declare const version = "abstract-signer/5.0.5";