+++ packages/providers/lib/etherscan-provider.js
@@ -125,14 +125,42 @@ function checkLogTag(blockTag) {
     return parseInt(blockTag.substring(2), 16);
 }
 var defaultApiKey = "9D13ZE7XSBTJ94N9BNJ2MA33VMAY2YPIRB";
-function checkGasError(error, transaction) {
+function checkError(method, error, transaction) {
+    // Get the message from any nested error structure
     var message = error.message;
-    if (error.code === logger_1.Logger.errors.SERVER_ERROR && error.error && typeof (error.error.message) === "string") {
-        message = error.error.message;
+    if (error.code === logger_1.Logger.errors.SERVER_ERROR) {
+        if (error.error && typeof (error.error.message) === "string") {
+            message = error.error.message;
+        }
+        else if (typeof (error.body) === "string") {
+            message = error.body;
+        }
+        else if (typeof (error.responseText) === "string") {
+            message = error.responseText;
+        }
+    }
+    message = (message || "").toLowerCase();
+    // "Insufficient funds. The account you tried to send transaction from does not have enough funds. Required 21464000000000 and got: 0"
+    if (message.match(/insufficient funds/)) {
+        logger.throwError("insufficient funds for intrinsic transaction cost", logger_1.Logger.errors.INSUFFICIENT_FUNDS, {
+            error: error, method: method, transaction: transaction
+        });
+    }
+    // "Transaction with the same hash was already imported."
+    if (message.match(/same hash was already imported|transaction nonce is too low/)) {
+        logger.throwError("nonce has already been used", logger_1.Logger.errors.NONCE_EXPIRED, {
+            error: error, method: method, transaction: transaction
+        });
+    }
+    // "Transaction gas price is too low. There is another transaction with same nonce in the queue. Try increasing the gas price or incrementing the nonce."
+    if (message.match(/another transaction with same nonce/)) {
+        logger.throwError("replacement fee too low", logger_1.Logger.errors.REPLACEMENT_UNDERPRICED, {
+            error: error, method: method, transaction: transaction
+        });
     }
     if (message.match(/execution failed due to an exception/)) {
         logger.throwError("cannot estimate gas; transaction may fail or may require manual gas limit", logger_1.Logger.errors.UNPREDICTABLE_GAS_LIMIT, {
-            error: error, transaction: transaction
+            error: error, method: method, transaction: transaction
         });
     }
     throw error;
@@ -271,21 +299,7 @@ var EtherscanProvider = /** @class */ (function (_super) {
                         url += "/api?module=proxy&action=eth_sendRawTransaction&hex=" + params.signedTransaction;
                         url += apiKey;
                         return [2 /*return*/, get(url).catch(function (error) {
-                                if (error.responseText) {
-                                    // "Insufficient funds. The account you tried to send transaction from does not have enough funds. Required 21464000000000 and got: 0"
-                                    if (error.responseText.toLowerCase().indexOf("insufficient funds") >= 0) {
-                                        logger.throwError("insufficient funds", logger_1.Logger.errors.INSUFFICIENT_FUNDS, {});
-                                    }
-                                    // "Transaction with the same hash was already imported."
-                                    if (error.responseText.indexOf("same hash was already imported") >= 0) {
-                                        logger.throwError("nonce has already been used", logger_1.Logger.errors.NONCE_EXPIRED, {});
-                                    }
-                                    // "Transaction gas price is too low. There is another transaction with same nonce in the queue. Try increasing the gas price or incrementing the nonce."
-                                    if (error.responseText.indexOf("another transaction with same nonce") >= 0) {
-                                        logger.throwError("replacement fee too low", logger_1.Logger.errors.REPLACEMENT_UNDERPRICED, {});
-                                    }
-                                }
-                                throw error;
+                                return checkError("sendTransaction", error, params.signedTransaction);
                             })];
                     case 8:
                         if (params.blockTag) {
@@ -326,7 +340,7 @@ var EtherscanProvider = /** @class */ (function (_super) {
                     case 13: return [2 /*return*/, _c.sent()];
                     case 14:
                         error_1 = _c.sent();
-                        return [2 /*return*/, checkGasError(error_1, params.transaction)];
+                        return [2 /*return*/, checkError("call", error_1, params.transaction)];
                     case 15:
                         transaction = getTransactionString(params.transaction);
                         if (transaction) {
@@ -341,7 +355,7 @@ var EtherscanProvider = /** @class */ (function (_super) {
                     case 17: return [2 /*return*/, _c.sent()];
                     case 18:
                         error_2 = _c.sent();
-                        return [2 /*return*/, checkGasError(error_2, params.transaction)];
+                        return [2 /*return*/, checkError("estimateGas", error_2, params.transaction)];
                     case 19:
                         url += "/api?module=logs&action=getLogs";
                         if (params.filter.fromBlock) {