+++ packages/web/lib.esm/index.js
@@ -29,7 +29,7 @@ function bodyify(value, type) {
         return value;
     }
     if (isBytesLike(value)) {
-        if (type && (type.split("/")[0] === "text" || type === "application/json")) {
+        if (type && (type.split("/")[0] === "text" || type.split(";")[0].trim() === "application/json")) {
             try {
                 return toUtf8String(value);
             }